add compatibility for moreMapIcons
